# docs/swagger.yaml
basePath: /
definitions:
...
  model.Platform:
    properties:
      createdAt:
        type: string
      deactivatedAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      domains:
        items:
          type: string
        type: array
      id:
        type: string
      ipAddresses:
        items:
          type: string
        type: array
      name:
        type: string
      organizationId:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.RequestPlatformCreate:
    properties:
      platformDescription:
        type: string
      platformName:
        type: string
    required:
    - platformName
    type: object
  model.RequestPlatformUpdate:
    properties:
      description:
        type: string
      domains:
        items:
          type: string
        type: array
      ipAddresses:
        items:
          type: string
        type: array
      platformName:
        type: string
    type: object
...
host: platform-api.string-api.xyz
info:
  contact:
    email: support@stringxyz.com
    name: Platform Management API Support
    url: http://string.xyz
  description: Platform Management API for managing Organizations, Platforms, API
    keys, etc.
  title: Platform Management API
  version: "1.0"
paths:
...
  /platforms:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Platform'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Get all platforms
      tags:
      - Platform
    post:
      consumes:
      - application/json
      parameters:
      - description: Platform Create Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RequestPlatformCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Platform'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Create platform
      tags:
      - Platform
  /platforms/{id}:
    get:
      parameters:
      - description: Platform ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Platform'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Get platform
      tags:
      - Platform
    patch:
      consumes:
      - application/json
      parameters:
      - description: Platform ID
        in: path
        name: id
        required: true
        type: string
      - description: Platform Update Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RequestPlatformUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Platform'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Update platform
      tags:
      - Platform
securityDefinitions:
  BearerAuth:
    type: apiKey
    in: header
    name: StringAdminJWT
  RefreshBearerAuth:
    type: apiKey
    in: header
    name: StringAdminRefreshToken
  JWTAuth:
    type: apiKey
    in: cookie
    name: StringAdminJWT
  RefreshTokenAuth:
    type: apiKey
    in: cookie
    name: StringAdminRefreshToken
swagger: "2.0"